+++ src/Gain/Holo/BackendCUDA.cs
@@ -23,63 +23,63 @@ namespace AUTD3Sharp.Gain.Holo
             Ptr.Item1 = IntPtr.Zero;
         }
 
-        internal override GainPtr Sdp(double[] foci, Amplitude[] amps, ulong size, double alpha, uint repeat, double lambda, EmissionConstraintWrap constraint)
+        internal override GainPtr Sdp(float[] foci, Amplitude[] amps, uint size, float alpha, uint repeat, float lambda, EmissionConstraintWrap constraint)
         {
             unsafe
             {
-                fixed (double* pf = foci)
+                fixed (float* pf = foci)
                 fixed (Amplitude* pa = amps)
                 {
-                    return NativeMethodsBackendCuda.AUTDGainHoloCUDASDP(Ptr, pf, (double*)pa, size, alpha, lambda, repeat, constraint);
+                    return NativeMethodsBackendCuda.AUTDGainHoloCUDASDP(Ptr, pf, (float*)pa, size, alpha, lambda, repeat, constraint);
                 }
             }
         }
 
-        internal override GainPtr Gs(double[] foci, Amplitude[] amps, ulong size, uint repeat, EmissionConstraintWrap constraint)
+        internal override GainPtr Gs(float[] foci, Amplitude[] amps, uint size, uint repeat, EmissionConstraintWrap constraint)
         {
             unsafe
             {
-                fixed (double* pf = foci)
+                fixed (float* pf = foci)
                 fixed (Amplitude* pa = amps)
                 {
-                    return NativeMethodsBackendCuda.AUTDGainHoloCUDAGS(Ptr, pf, (double*)pa, size, repeat, constraint);
+                    return NativeMethodsBackendCuda.AUTDGainHoloCUDAGS(Ptr, pf, (float*)pa, size, repeat, constraint);
                 }
             }
         }
 
-        internal override GainPtr Gspat(double[] foci, Amplitude[] amps, ulong size, uint repeat, EmissionConstraintWrap constraint)
+        internal override GainPtr Gspat(float[] foci, Amplitude[] amps, uint size, uint repeat, EmissionConstraintWrap constraint)
         {
             unsafe
             {
-                fixed (double* pf = foci)
+                fixed (float* pf = foci)
                 fixed (Amplitude* pa = amps)
                 {
-                    return NativeMethodsBackendCuda.AUTDGainHoloCUDAGSPAT(Ptr, pf, (double*)pa, size, repeat, constraint);
+                    return NativeMethodsBackendCuda.AUTDGainHoloCUDAGSPAT(Ptr, pf, (float*)pa, size, repeat, constraint);
                 }
             }
         }
 
-        internal override GainPtr Naive(double[] foci, Amplitude[] amps, ulong size, EmissionConstraintWrap constraint)
+        internal override GainPtr Naive(float[] foci, Amplitude[] amps, uint size, EmissionConstraintWrap constraint)
         {
             unsafe
             {
-                fixed (double* pf = foci)
+                fixed (float* pf = foci)
                 fixed (Amplitude* pa = amps)
                 {
-                    return NativeMethodsBackendCuda.AUTDGainHoloCUDANaive(Ptr, pf, (double*)pa, size, constraint);
+                    return NativeMethodsBackendCuda.AUTDGainHoloCUDANaive(Ptr, pf, (float*)pa, size, constraint);
                 }
             }
         }
 
-        internal override GainPtr Lm(double[] foci, Amplitude[] amps, ulong size, double eps1, double eps2, double tau, uint kMax, double[] initial, EmissionConstraintWrap constraint)
+        internal override GainPtr Lm(float[] foci, Amplitude[] amps, uint size, float eps1, float eps2, float tau, uint kMax, float[] initial, EmissionConstraintWrap constraint)
         {
             unsafe
             {
-                fixed (double* pf = foci)
+                fixed (float* pf = foci)
                 fixed (Amplitude* pa = amps)
-                fixed (double* pInitial = initial)
+                fixed (float* pInitial = initial)
                 {
-                    return NativeMethodsBackendCuda.AUTDGainHoloCUDALM(Ptr, pf, (double*)pa, size, eps1, eps2, tau, kMax, constraint, pInitial, (ulong)initial.Length);
+                    return NativeMethodsBackendCuda.AUTDGainHoloCUDALM(Ptr, pf, (float*)pa, size, eps1, eps2, tau, kMax, constraint, pInitial, (ulong)initial.Length);
                 }
             }
         }
