+++ src/Gain/Holo/Constraint.cs
@@ -12,7 +12,7 @@ namespace AUTD3Sharp.Gain.Holo
         public static EmissionConstraintWrap Normalize => NativeMethodsGainHolo.AUTDGainHoloConstraintNormalize();
         public static EmissionConstraintWrap Uniform(EmitIntensity value) => NativeMethodsGainHolo.AUTDGainHoloConstraintUniform(value.Value);
         public static EmissionConstraintWrap Clamp(EmitIntensity min, EmitIntensity max) => NativeMethodsGainHolo.AUTDGainHoloConstraintClamp(min.Value, max.Value);
-        public static EmissionConstraintWrap Multiply(double value) => NativeMethodsGainHolo.AUTDGainHoloConstraintMultiply(value);
+        public static EmissionConstraintWrap Multiply(float value) => NativeMethodsGainHolo.AUTDGainHoloConstraintMultiply(value);
     }
 }
